flask-cicd-app/requirements.txt
flask==3.0.0
pytest==7.4.3
mysql-connector-python
prometheus-client
